refactor(api): define Mesh family as Go structs
Mesh, MeshInsight and MeshOverview convert together because they cannot
convert apart: MeshOverview embeds both of the others, and a Go struct
holding a protobuf message does not serialize to the bytes jsonpb wrote.

The Mesh spec is the stored one and carries a single live field, the rest
having been reserved over time. The insight is recomputed and the overview
is assembled for the API, so neither strands operator data.

sameInsight compares the bytes each insight would be stored as. An insight
is built with an initialised resources map while one read back leaves it
nil, and proto.Equal treated those as equal, so a nil-sensitive comparison
would write on every resync.

MatchProto compares Go struct specs as well, since the matcher is handed
whichever of the two a converted resource now holds.
