--- api/mesh/v1alpha1/mesh.proto
@@ -1,48 +1,0 @@
-syntax = "proto3";
-
-package kuma.mesh.v1alpha1;
-
-import "api/mesh/options.proto";
-
-option go_package = "github.com/kumahq/kuma/v3/api/mesh/v1alpha1";
-
-// Mesh defines configuration of a single mesh.
-message Mesh {
-  option (kuma.mesh.resource).name = "MeshResource";
-  option (kuma.mesh.resource).type = "Mesh";
-  option (kuma.mesh.resource).global = true;
-  option (kuma.mesh.resource).package = "mesh";
-  option (kuma.mesh.resource).kds = "model.GlobalToZonesFlag";
-  option (kuma.mesh.resource).ws.name = "mesh";
-  option (kuma.mesh.resource).ws.plural = "meshes";
-  option (kuma.mesh.resource).has_insights = true;
-  option (kuma.mesh.resource).short_name = "m";
-
-  reserved 1;
-  reserved "mtls";
-
-  reserved 2;
-  reserved "tracing";
-
-  reserved 3;
-  reserved "logging";
-
-  reserved 4;
-  reserved "metrics";
-
-  reserved 5;
-  reserved "networking";
-
-  reserved 6;
-  reserved "routing";
-
-  reserved 7;
-  reserved "constraints";
-
-  // List of policies to skip creating by default when the mesh is created.
-  // e.g. TrafficPermission, MeshRetry, etc. An '*' can be used to skip all
-  // policies.
-  repeated string skipCreatingInitialPolicies = 8;
-
-  reserved 9; // formerly meshServices
-}
