--- api/mesh/v1alpha1/mesh_insight.proto
@@ -1,73 +1,0 @@
-syntax = "proto3";
-
-package kuma.mesh.v1alpha1;
-
-import "api/mesh/options.proto";
-
-option go_package = "github.com/kumahq/kuma/v3/api/mesh/v1alpha1";
-
-// MeshInsight defines the observed state of a Mesh.
-message MeshInsight {
-  option (kuma.mesh.resource).name = "MeshInsightResource";
-  option (kuma.mesh.resource).type = "MeshInsight";
-  option (kuma.mesh.resource).global = true;
-  option (kuma.mesh.resource).package = "mesh";
-  option (kuma.mesh.resource).ws.name = "mesh-insight";
-  option (kuma.mesh.resource).ws.read_only = true;
-  // MeshInsights are only managed globally
-  option (kuma.mesh.resource).kds = "model.ProvidedByGlobalFlag";
-  option (kuma.mesh.resource).policy_matching_exempt = true;
-
-  reserved 1; // formerly last_sync
-  reserved 3; // formerly policies
-  reserved "policies";
-
-  // DataplaneStat defines statistic specifically for Dataplane
-  message DataplaneStat {
-    uint32 total = 1;
-    uint32 online = 2;
-    uint32 offline = 3;
-    uint32 partially_degraded = 4;
-  }
-  DataplaneStat dataplanes = 2;
-
-  // DpVersions defines statistics grouped by dataplane versions
-  message DpVersions {
-    // Dataplane stats grouped by KumaDP version
-    map<string, DataplaneStat> kumaDp = 1;
-
-    // Dataplane stats grouped by Envoy version
-    map<string, DataplaneStat> envoy = 2;
-  }
-  DpVersions dpVersions = 4;
-
-  message MTLS {
-    // Dataplanes grouped by issued backends.
-    map<string, DataplaneStat> issuedBackends = 1;
-    // Dataplanes grouped by supported backends.
-    map<string, DataplaneStat> supportedBackends = 2;
-  }
-
-  // mTLS statistics
-  MTLS mTLS = 5;
-
-  reserved 6; // formerly services
-  reserved "services";
-
-  // DataplanesByType defines statistics splitted by dataplane types
-  message DataplanesByType {
-    DataplaneStat standard = 1;
-    DataplaneStat gateway = 2;
-
-    reserved 3; // formerly gatewayBuiltin, removed with the built-in gateway
-    reserved "gatewayBuiltin";
-
-    DataplaneStat gatewayDelegated = 4;
-  }
-  DataplanesByType dataplanesByType = 7;
-
-  message ResourceStat {
-    uint32 total = 1;
-  }
-  map<string, ResourceStat> resources = 8;
-}
